refactor(plugin-code): move SpecView to components
SpecView renders MDL from a plain string and uses no ECHO/identity/data
hooks (only theme), so it is presentational and belongs in components/,
not co-located inside the ECHO-bound SpecArticle container. This also
removes the deep barrel re-export that the co-location forced.
