+++ packages/plugins/plugin-code/src/components/SpecView/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2025 DXOS.org
+//
+
+export * from './SpecView';
