+++ packages/plugins/plugin-code/src/components/index.ts
@@ -5,3 +5,4 @@
 export * from './BuildOutput';
 export * from './CodeToolbar';
 export * from './FileTree';
+export * from './SpecView';
