+++ packages/plugins/plugin-code/src/containers/index.ts
@@ -7,4 +7,3 @@ import { type ComponentType, lazy } from 'react';
 export const CodeArticle: ComponentType<any> = lazy(() => import('./CodeArticle'));
 export const CodeSettings: ComponentType<any> = lazy(() => import('./CodeSettings'));
 export const SpecArticle: ComponentType<any> = lazy(() => import('./SpecArticle'));
-export { SpecView } from './SpecArticle/SpecArticle';
